brought the old survey module back and moved getFillStateForAllCols back out of the janno module
